Modernize Stage component
- Use typescript
- Add type information for pipeline rule
- Split component into two smaller ones
- Replace createReactClass with functional component
- Fix linter issues
